+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 ### v0.2.0 - 2024-01-16
 
 - remove all code related to seed indexing and querying and minimize the dependencies.
+- Mask() result: only use the last 1 bit, rather than 2, for storing the strand information